requirements.txt
adal==1.2.7
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
argcomplete==3.6.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
azure-common==1.1.28
azure-core==1.33.0
azure-graphrbac==0.61.2
azure-mgmt-authorization==4.0.0
azure-mgmt-containerregistry==10.3.0
azure-mgmt-core==1.5.0
azure-mgmt-keyvault==10.3.1
azure-mgmt-network==28.1.0
azure-mgmt-resource==23.3.0
azure-mgmt-storage==22.2.0
azureml-core==1.60.0
babel==2.17.0
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==4.3.0
beautifulsoup4==4.13.4
bleach==6.2.0
blinker==1.9.0
boto3==1.38.1
botocore==1.38.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
comm==0.2.2
contextlib2==21.6.0
contourpy==1.3.2
cryptography==44.0.2
cycler==0.12.1
databricks-sdk==0.50.0
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
distlib==0.3.9
docker==7.1.0
durationpy==0.9
executing==2.2.0
fastapi>=0.104.0
fastjsonschema==2.21.1
filelock==3.18.0
Flask==3.1.0
fonttools==4.57.0
fqdn==1.5.1
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httpx>=0.25.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.9.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
json5==0.12.0
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupytext==1.17.0
kiwisolver==1.4.8
knack==0.12.0
kubernetes==32.0.1
Mako==1.3.10
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib>=3.5.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.1.3
mlflow>=2.9.0
mlflow-skinny==2.22.0
msal==1.32.0
msal-extensions==1.3.1
msrest==0.7.1
msrestazure==0.6.4.post1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
notebook==7.4.1
notebook_shim==0.2.4
numpy>=1.24.0
oauthlib==3.2.2
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-semantic-conventions==0.53b1
overrides==7.7.0
packaging==24.2
pandas>=2.0.0
pandocfilters==1.5.1
paramiko==3.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
pkginfo==1.12.1.2
platformdirs==4.3.7
prometheus_client==0.21.1
prometheus_flask_exporter==0.23.2
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic>=2.0.0
pydantic_core==2.33.1
Pygments==2.19.1
PyJWT==2.10.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pysftp==0.2.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-jose[cryptography]>=3.3.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
referencing==0.36.2
requests==2.32.3
requests-auth-aws-sigv4==0.7
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.24.0
rsa==4.9.1
s3transfer==0.12.0
scikit-learn>=1.3.0
scipy==1.15.2
SecretStorage==3.3.3
Send2Trash==1.8.3
setuptools==79.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
tabulate==0.9.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.4.0
uvicorn>=0.23.0
virtualenv==20.30.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
seaborn>=0.11.0
python-dotenv>=1.0.0
supabase>=1.0.3
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
structlog>=23.2.0
redis>=4.6.0
fastapi-cache2>=0.2.1
fastapi-cache2[redis]>=0.2.1
xgboost>=2.0.0
psycopg2-binary
